Add files via upload
Added _header and _footer partials (.logo is in header), mobile versions.  Added several images.  Changed the position of the "Moksha" heading to header in index.html.  Made a few small changes to index.html.  Commented out the CSS animation for the banner image text because it was clipping off the page.  Added the _contact partial and almost completed the contact page for mobile.